+++ KanjiPractice/congratulators.json
@@ -1,0 +1,17 @@
+{
+    "1": [
+        "Wow, senpai! You aced it on the first try! Miku is impressed!",
+        "Amazing job! Miku knew you could get it right from the start!",
+        "Woah! You're a quick learner. Miku's virtual applause for you!"
+    ],
+    "2": [
+        "Well done on the second try! Miku knew you just needed a warm-up.",
+        "Nice recovery! Miku believes in your persistence and success!",
+        "Correct! Miku is impressed with your kanji skills!"
+    ],
+    "3": [
+        "Persistence pays off! You got it on the third try. Miku is proud of you!",
+        "You did it! Miku knew you had it in you!",
+        "Great job! You nailed it, senpai!"
+    ]
+}
