+++ KanjiPractice/encouragers.json
@@ -1,0 +1,17 @@
+{
+    "1": [
+        "Close, but not quite! Miku believes in your ability to learn.",
+        "Oopsie! Not quite there, but you've got more chances to get it right!",
+        "No worries, senpai! Everyone makes mistakes. Give it another shot!"
+    ],
+    "2": [
+        "Not quite, but don't give up! Miku is cheering you on for the final try.",
+        "Hmm...",
+        "Are you sure about that?"
+    ],
+    "3": [
+        "It's okay, senpai! Learning is a journey. Let's try another question!",
+        "No worries! Even if this one got away, there are more chances ahead.",
+        "That's alright! Remember, it's all about progress. Miku is proud of your effort!"
+    ]
+}
